requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
adrf==0.1.2

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Redis
redis>=5.0,<6.0

# Utilities
nanoid>=2.0,<3.0
Faker>=22.0,<23.0
uvicorn[standard]==0.27.0

# API Documentation
drf-spectacular>=0.27,<1.0


# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
httpx>=0.26,<1.0